requirements.txt
streamlit
pandas
python-dotenv
crewai
groq
scikit-learn
matplotlib
seaborn
numpy
